gaussian_splatting: preserve full 64-bit asset identity through renderer remap
Codex review on PR #409 found the asset_id 64->32 truncation fix incomplete at
the renderer boundary. Keying SharedWorld::asset_records on the full ObjectID
fixed host-side aliasing, but the renderer remap still collapsed colliding ids:

  - InstanceAssetRegistration::asset_id was uint32_t
  - PublishedInstanceAssetRemap::asset_to_dense_id was HashMap<uint32_t,uint32_t>
  - update_instance_buffer() looked the dense slot up by the truncated 32-bit
    InstanceDataGPU::ids[0] GPU tag

So two assets whose ObjectIDs differ only above the low 32 bits resolved to the
SAME dense atlas slot and rendered as one asset.

Fix: keep a collision-free 64-bit submission identity through the whole remap
path, while the GPU-facing dense slot stays 32-bit (it is a sequential,
collision-free index):

  - Widen InstanceAssetRegistration::asset_id and the asset_to_dense_id map KEY
    to uint64_t (host submission identity); the dense_id VALUE stays uint32_t.
  - collect_*_for_renderer no longer truncates the asset_records key.
  - build_instance_buffer_for_renderer emits a parallel LocalVector<uint64_t> of
    submission ids; update_instance_buffer uses it as the collision-free lookup
    key, then writes the resolved DENSE id into ids[0]. ids[0] is now treated as
    a transient/output-only GPU field, never the authoritative host identity.
  - Resident publisher and streaming orchestrator thread the parallel submission
    ids through; the streaming registry stays internally 32-bit (a separate,
    GPU-atlas-coupled follow-up) but the published remap key is now 64-bit, so the
    resident path is collision-free.

GPU struct byte layout is unchanged: InstanceDataGPU / PackedGaussian and their
GLSL mirrors are untouched, and only the 32-bit dense id reaches the GPU.
gaussian-layout-check passes.

Tests: headless remap-distinctness assertion added to the asset_id collision
suite, plus a RequiresGPU end-to-end case asserting two low-32-bit-colliding
submission ids resolve to distinct dense slots through update_instance_buffer.
RequiresGPU manifest snapshot refreshed (109->110).

+++ modules/gaussian_splatting/renderer/render_types/render_pipeline_io_types.h
@@ -167,7 +167,12 @@ inline const char *instance_backend_policy_to_string(InstanceBackendPolicy p_pol
 }
 
 struct PublishedInstanceAssetRemap {
-	HashMap<uint32_t, uint32_t> asset_to_dense_id;
+	// Key: full 64-bit host submission asset identity (asset Node3D ObjectID).
+	// Value: sequential dense atlas slot id (collision-free, fits uint32_t and is
+	// what lands in the GPU InstanceDataGPU::ids[0] field). The key must stay
+	// 64-bit so two assets whose ObjectIDs collide in the low 32 bits map to
+	// distinct dense slots instead of aliasing the same rendered asset.
+	HashMap<uint64_t, uint32_t> asset_to_dense_id;
 	uint64_t generation = 0;
 	bool valid = false;
